Use 'build' when referencing the root part of the config.
For the sake of consistency use the name of the node itself, 'build', when
requesting the entire tree or a config key in the root or 'build' section,
such as sites_subdir or make. Don't use 'all', which was perhaps accurate
but a bit confusing.

Calls to config() probably still need refining at this point.